https://jira.aplana.com/browse/SBRFNDFL-5501 Настройки подразделений. Сортировки поправил

+++ db/changelog/templates-3.1/ndfl/report_2ndfl_1/v2016/report_2ndfl.groovy
@@ -1402,7 +1402,7 @@ class Report2Ndfl extends AbstractScriptClass {
 
     final int RNU_NDFL_DECLARATION_TYPE = 101
 
-    Map<Long, List<Map<String, RefBookValue>>> departmentConfigsCache = [:]
+    Map<Integer, List<Map<String, RefBookValue>>> departmentConfigsCache = [:]
 
     def createForm() {
         try {
